refactor: 팀 보따리 편집 화면의 EmptyView 공용 컴포넌트로 분리
