+++ android/Bottari/presentation/src/main/java/com/bottari/presentation/compose/edit/team/shared/TeamSharedScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.bottari.presentation.compose.edit.team.TeamChecklistEditScreen
+import com.bottari.presentation.compose.edit.team.component.TeamChecklistEditScreen
 import com.bottari.presentation.model.bottari.BottariItemUiModel
 import com.bottari.presentation.model.bottari.personal.BottariItemTypeUiModel
